move event host to separate file
